+++ Assets/PotatoUtil/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2022-11-04
+### Added
+- `ReadOnlyDatabase` that can be created based on a `Database`
+- `AsReadOnly` function on `Database`
+
 ## [3.2.0] - 2022-10-29
 ### Added
 - `FNVHash` can now be created from a `StringBuilder`